autogen-security-advisor/requirements.txt
autogen
langchain
openai
faiss-cpu
python-dotenv
requests
tiktoken
langchain_community
